+++ src/components/FHIRendpoint.js
@@ -1,19 +1,13 @@
 import { useState } from 'react';
-import { Icon } from '@iconify/react';
 import Advanced from './AdvancedOptions';
 
 function Endpoint() {
   const [link, setLink] = useState('');
-  const [showAdvancedOptions, setShowAdvancedOptions] = useState(false);
   const [buttonHover, setButtonHover] = useState(false);
   const [buttonClick, setButtonClick] = useState(false);
 
   const handleSubmit = (event) => {
     event.preventDefault();
-  };
-
-  const handleAdvancedOptionsClick = () => {
-    setShowAdvancedOptions((prev) => !prev);
   };
 
   const handleButtonHover = () => {
@@ -39,7 +33,7 @@ function Endpoint() {
             type="text"
             value={link}
             onChange={(e) => setLink(e.target.value)}
-            placeholder="https://api.examplePath.org/RTTD/open"
+            placeholder="http://example.org/"
             className="w-full rounded-lg h-10 shadow-widgit pl-2 pr-10"
           />
         </label>
@@ -55,30 +49,7 @@ function Endpoint() {
           Connect
         </button>
       </div>
-      <div
-        className="flex items-center cursor-pointer pt-1 font"
-        role="button"
-        tabIndex={0}
-        onClick={handleAdvancedOptionsClick}
-        onKeyDown={(event) => {
-          if (event.key === 'Enter' || event.key === ' ') {
-            handleAdvancedOptionsClick();
-          }
-        }}
-      >
-        <p>Advanced Options</p>
-        <Icon
-          icon="bi:caret-down-fill"
-          className="ml-1 mt-1"
-          height="12"
-          rotate={showAdvancedOptions ? '0deg' : '270deg'}
-        />
-      </div>
-      {showAdvancedOptions && (
-        <div>
-          <Advanced />
-        </div>
-      )}
+      <Advanced />
     </form>
   );
 }
